Refactoring, moves test runs in a separate class
