+++ src/roadgraph/SimpleMapGraph.java
@@ -199,7 +199,7 @@ public class SimpleMapGraph extends MapGraph{
 			//System.out.println("\ntoExplore - " + Arrays.toString(toExplore.toArray()));
 			//System.out.println("visited - " + Arrays.toString(visited.toArray()));			
 			MapNode currentNode = toExplore.remove();
-			System.out.println("Dijkstra visiting - " + currentNode);	
+			//System.out.println("Dijkstra visiting - " + currentNode);	
 			visitedCount++;
 
 			if (!visited.contains(currentNode)) {
@@ -207,7 +207,8 @@ public class SimpleMapGraph extends MapGraph{
 				visited.add(currentNode);
 				if(currentNode.equals(endNode)) {
 					found = true;
-					System.out.println("\nDijkstra - Total nodes visited - "+ visitedCount);
+					System.out.println("\nDijkstra");
+					System.out.println("\tNodes visited - " + visitedCount);
 					break;
 				}
 				HashMap<MapNode, Double> neighborDistanceMap = currentNode.getNeighborDistanceMap();
@@ -310,7 +311,7 @@ public class SimpleMapGraph extends MapGraph{
 			//System.out.println("\ntoExplore - " + toExplore);
 			//System.out.println("visited - " + visited);			
 			MapNode currentNode = toExplore.remove();
-			System.out.println("AStar visiting - " + currentNode);
+			//System.out.println("AStar visiting - " + currentNode);
 			visitedCount++;
 			
 			if (!visited.contains(currentNode)) {
@@ -318,7 +319,8 @@ public class SimpleMapGraph extends MapGraph{
 				visited.add(currentNode);
 				if(currentNode.equals(endNode)) {
 					found = true;
-					System.out.println("\nAStar - Total nodes visited - "+ visitedCount);
+					System.out.println("\nAStar");
+					System.out.println("\tNodes visited - " + visitedCount);
 					break;
 				}
 				HashMap<MapNode, Double> neighborDistanceMap = currentNode.getNeighborDistanceMap();
@@ -350,176 +352,5 @@ public class SimpleMapGraph extends MapGraph{
 
 		// Construct & return path
 		return constructPathFromParentMap(startNode, endNode, parentMap);
-	}
-
-	static void simpleTest() {
-		GeographicPoint testStart;
-		GeographicPoint testEnd;
-		List<GeographicPoint> testroute;
-		List<GeographicPoint> testroute2;
-		
-		MapGraph simpleTestMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/testdata/simpletest.map", simpleTestMap);
-		System.out.println("Map - ");
-		simpleTestMap.printGraph();
-		
-		testStart = new GeographicPoint(1.0, 1.0);
-		testEnd = new GeographicPoint(8.0, -1.0);
-		
-		System.out.println("Test 1 using simpletest: Dijkstra should be 9 and AStar should be 5");
-		testroute = simpleTestMap.dijkstra(testStart,testEnd);
-		System.out.println("Dijkstra result - " + testroute);;
-		
-		System.out.println("----------------------------------------------------");
-		simpleTestMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/testdata/simpletest.map", simpleTestMap);
-		testroute2 = simpleTestMap.aStarSearch(testStart,testEnd);
-		System.out.println("AStar result- " + testroute2);		
-	}
-	
-	static void utcTest() {
-		GeographicPoint testStart;
-		GeographicPoint testEnd;
-		List<GeographicPoint> testroute;
-		List<GeographicPoint> testroute2;
-		
-		// A very simple test using real data
-		testStart = new GeographicPoint(32.869423, -117.220917);
-		testEnd = new GeographicPoint(32.869255, -117.216927);
-		System.out.println("\n\nTest 2 using utc: Dijkstra should be 13 and AStar should be 5");
-		
-		SimpleMapGraph testMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/maps/utc.map", testMap);
-		testMap.printGraph();
-		testroute = testMap.dijkstra(testStart,testEnd);
-		System.out.println("Dijkstra result - " + testroute);;
-		
-		System.out.println("----------------------------------------------------");
-		testMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/maps/utc.map", testMap);
-		testroute2 = testMap.aStarSearch(testStart,testEnd);
-		System.out.println("AStar result- " + testroute2);
-
-
-	}
-	static void utcTestComplexDijkstra() {
-		GeographicPoint testStart;
-		GeographicPoint testEnd;
-		List<GeographicPoint> testroute;
-		
-		// A very simple test using real data
-		testStart = new GeographicPoint(32.8674388, -117.2190213);
-		testEnd = new GeographicPoint(32.8697828, -117.2244506);
-		System.out.println("Test 3 using utc: Dijkstra should be 37");
-		
-		SimpleMapGraph testMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/maps/utc.map", testMap);	
-		testroute = testMap.dijkstra(testStart,testEnd);
-		System.out.println("Dijkstra result - " + testroute);;
-		
-	}
-	static void utcTestComplexAStar() {
-		GeographicPoint testStart;
-		GeographicPoint testEnd;
-		List<GeographicPoint> testroute2;
-		
-		// A very simple test using real data
-		testStart = new GeographicPoint(32.8674388, -117.2190213);
-		testEnd = new GeographicPoint(32.8697828, -117.2244506);
-		System.out.println("Test 3 using utc: AStar should be 10");
-		
-		SimpleMapGraph testMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/maps/utc.map", testMap);	
-		testroute2 = testMap.aStarSearch(testStart,testEnd);
-		System.out.println("AStar result- " + testroute2);
-		
-	}
-	static void utcTestQuiz() {
-		GeographicPoint testStart;
-		GeographicPoint testEnd;
-		List<GeographicPoint> testroute;
-		List<GeographicPoint> testroute2;
-		
-		// A very simple test using real data
-		testStart = new GeographicPoint(32.8648772, -117.2254046);
-		testEnd = new GeographicPoint(32.8660691, -117.217393);
-		
-		SimpleMapGraph testMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/maps/utc.map", testMap);
-		//testMap.printGraph();
-		testroute = testMap.dijkstra(testStart,testEnd);
-		System.out.println("Dijkstra result - " + testroute);;
-		
-		System.out.println("----------------------------------------------------");
-		testMap = new SimpleMapGraph();
-		GraphLoader.loadRoadMap("data/maps/utc.map", testMap);
-		testroute2 = testMap.aStarSearch(testStart,testEnd);
-		System.out.println("AStar result- " + testroute2);
-		
-	}
-	static void testDistanceFromStartComparison() {
-		MapNode node1 = new MapNode(new GeographicPoint(1.2, -1.2));
-		node1.setDistanceFromStart(3.0);
-		System.out.println(node1);
-		
-		MapNode node2 = new MapNode(new GeographicPoint(1.4, -1.4));
-		node2.setDistanceFromStart(2.1);
-		System.out.println(node2);
-
-		int result = node1.compareTo(node2);
-		System.out.println(result);
-		if (result < 0)
-			System.out.println("Success - node1 is farther to start than node2");
-		else
-			System.out.println("Failure");
-
-	}
-	static void testPriorityQueueBasedOnDistanceFromStart() {
-		MapNode node1 = new MapNode(new GeographicPoint(32.8674388, -117.2190213));
-		node1.setDistanceFromStart(1.12);
-		System.out.println(node1);
-		
-		MapNode node2 = new MapNode(new GeographicPoint(32.8697828, -117.2244506));
-		node2.setDistanceFromStart(1.134);
-		System.out.println(node2);
-
-		MapNode node3 = new MapNode(new GeographicPoint(32.8697828, -117.2244506));
-		node3.setDistanceFromStart(1.135);
-		System.out.println(node3);
-		
-		Queue<MapNode> priorityQueue = new PriorityQueue<MapNode>(new DistanceFromStartComparator());
-		Queue<MapNode> simpleQueue = new LinkedList<MapNode>();
-		
-		priorityQueue.add(node3);
-		simpleQueue.add(node3);
-		priorityQueue.add(node2);
-		simpleQueue.add(node2);
-		priorityQueue.add(node1);
-		simpleQueue.add(node1);
-		
-		System.out.println("Simple queue - ");
-		while(!simpleQueue.isEmpty())
-			System.out.println(simpleQueue.remove());
-
-		System.out.println("Priority queue - ");
-		while(!priorityQueue.isEmpty())
-			System.out.println(priorityQueue.remove());
-
-	}
-	public static void main(String[] args)
-	{
-		//testDistanceFromStartComparison();
-		//testPriorityQueueBasedOnDistanceFromStart();
-		
-		//GraphLoader.createIntersectionsFile("data/maps/utc.map",
-        //        "data/intersections/utc.intersections");
-		
-
-		simpleTest();
-		//utcTest();
-		//utcTestComplexDijkstra();
-		//utcTestComplexAStar();
-		//utcTestQuiz();
-	}
-	
+	}	
 }
